requirements.txt
Flask
google-generativeai
numpy
openpyxl
opencv-python
pandas
pdf-redactor
pdfplumber
Pillow
presidio-analyzer
presidio-anonymizer
PyMuPDF
PyPDF2
pytesseract
python-docx
python-pptx
spacy
streamlit
ultralytics